data/tools/attack_tools/pyproject.toml
[project]
name = "attack-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "docx2pdf>=0.1.8",
    "fastmcp>=2.12.2",
    "httpx>=0.28.1",
    "mcp[cli]>=1.12.2",
    "mem0ai>=0.1.117",
    "msoffcrypto-tool>=5.4.2",
    "python-docx>=1.2.0",
]
